Inline RspecGenerator
